# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ProductCategory = typeof ProductCategory[keyof typeof ProductCategory];


export const ProductCategory = {
  saree: 'saree',
  kurti: 'kurti',
  puja: 'puja',
  decor: 'decor',
} as const;

export type StockStatus = typeof StockStatus[keyof typeof StockStatus];


export const StockStatus = {
  in_stock: 'in_stock',
  made_to_order: 'made_to_order',
} as const;

export interface ProductVariant {
  label: string;
  /** @nullable */
  size?: string | null;
  /** @nullable */
  color?: string | null;
}

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  price: number;
  description: string;
  photos: string[];
  variants: ProductVariant[];
  stockStatus: StockStatus;
  active: boolean;
  createdAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  category: ProductCategory;
  /** @minimum 0 */
  price: number;
  description: string;
  photos?: string[];
  variants?: ProductVariant[];
  stockStatus?: StockStatus;
  active?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ProductCategory;
  /** @minimum 0 */
  price?: number;
  description?: string;
  photos?: string[];
  variants?: ProductVariant[];
  stockStatus?: StockStatus;
  active?: boolean;
}

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  upi: 'upi',
  cod: 'cod',
} as const;

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  received: 'received',
  packed: 'packed',
  shipped: 'shipped',
} as const;

export interface OrderItem {
  productId: number;
  productName: string;
  /** @nullable */
  variantLabel?: string | null;
  /** @minimum 1 */
  quantity: number;
  price: number;
}

export interface OrderItemInput {
  productId: number;
  /** @nullable */
  variantLabel?: string | null;
  /** @minimum 1 */
  quantity: number;
}

export interface Order {
  id: number;
  customerName: string;
  phone: string;
  address: string;
  items: OrderItem[];
  subtotal: number;
  deliveryFee: number;
  total: number;
  paymentMethod: PaymentMethod;
  status: OrderStatus;
  /** @nullable */
  estimatedDate?: string | null;
  createdAt: string;
}

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minItems 1 */
  items: OrderItemInput[];
  paymentMethod: PaymentMethod;
}

export interface OrderStatusUpdate {
  status: OrderStatus;
}

export interface OrdersSummary {
  received: number;
  packed: number;
  shipped: number;
  total: number;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AdminSession {
  authenticated: boolean;
}

export interface UploadUrlInput {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListProductsParams = {
category?: ProductCategory;
search?: string;
includeInactive?: boolean;
};

export type ListOrdersParams = {
status?: OrderStatus;
};