# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: orders
    description: Order operations
  - name: admin
    description: Admin auth operations
  - name: storage
    description: Object storage operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      description: List products, optionally filtered by category. By default only active products are returned unless includeInactive is true (admin use).
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProductCategory"
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: includeInactive
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin only), newest first
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/OrderStatus"
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/summary:
    get:
      operationId: getOrdersSummary
      tags: [orders]
      summary: Order counts by status (admin only) for dashboard badges
      responses:
        "200":
          description: Summary counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersSummary"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by id (used for order confirmation & tracking, public via id)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Log in as admin (mom)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Log out admin session
      responses:
        "204":
          description: Logged out

  /admin/session:
    get:
      operationId: getAdminSession
      tags: [admin]
      summary: Check current admin session status
      responses:
        "200":
          description: Session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlInput"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    ProductCategory:
      type: string
      enum: [saree, kurti, puja, decor]

    StockStatus:
      type: string
      enum: [in_stock, made_to_order]

    ProductVariant:
      type: object
      required: [label]
      properties:
        label:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]

    Product:
      type: object
      required:
        - id
        - name
        - category
        - price
        - description
        - photos
        - variants
        - stockStatus
        - active
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          $ref: "#/components/schemas/ProductCategory"
        price:
          type: number
        description:
          type: string
        photos:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        stockStatus:
          $ref: "#/components/schemas/StockStatus"
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time

    ProductInput:
      type: object
      required: [name, category, price, description]
      properties:
        name:
          type: string
          minLength: 1
        category:
          $ref: "#/components/schemas/ProductCategory"
        price:
          type: number
          minimum: 0
        description:
          type: string
        photos:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        stockStatus:
          $ref: "#/components/schemas/StockStatus"
        active:
          type: boolean

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          $ref: "#/components/schemas/ProductCategory"
        price:
          type: number
          minimum: 0
        description:
          type: string
        photos:
          type: array
          items:
            type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        stockStatus:
          $ref: "#/components/schemas/StockStatus"
        active:
          type: boolean

    PaymentMethod:
      type: string
      enum: [upi, cod]

    OrderStatus:
      type: string
      enum: [received, packed, shipped]

    OrderItem:
      type: object
      required: [productId, productName, quantity, price]
      properties:
        productId:
          type: integer
        productName:
          type: string
        variantLabel:
          type: ["string", "null"]
        quantity:
          type: integer
          minimum: 1
        price:
          type: number

    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        variantLabel:
          type: ["string", "null"]
        quantity:
          type: integer
          minimum: 1

    Order:
      type: object
      required:
        - id
        - customerName
        - phone
        - address
        - items
        - subtotal
        - deliveryFee
        - total
        - paymentMethod
        - status
        - createdAt
      properties:
        id:
          type: integer
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotal:
          type: number
        deliveryFee:
          type: number
        total:
          type: number
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        status:
          $ref: "#/components/schemas/OrderStatus"
        estimatedDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time

    OrderInput:
      type: object
      required: [customerName, phone, address, items, paymentMethod]
      properties:
        customerName:
          type: string
          minLength: 1
        phone:
          type: string
          minLength: 1
        address:
          type: string
          minLength: 1
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/OrderItemInput"
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"

    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"

    OrdersSummary:
      type: object
      required: [received, packed, shipped, total]
      properties:
        received:
          type: integer
        packed:
          type: integer
        shipped:
          type: integer
        total:
          type: integer

    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
          minLength: 1

    AdminSession:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean

    UploadUrlInput:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
        size:
          type: number
        contentType:
          type: string

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string