# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';
import type { PaymentMethod } from './paymentMethod';

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  address: string;
  /** @minItems 1 */
  items: OrderItemInput[];
  paymentMethod: PaymentMethod;
}