# lib/api-zod/src/generated/types/orderItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderItem {
  productId: number;
  productName: string;
  /** @nullable */
  variantLabel?: string | null;
  /** @minimum 1 */
  quantity: number;
  price: number;
}