# lib/api-zod/src/generated/types/orderStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  received: 'received',
  packed: 'packed',
  shipped: 'shipped',
} as const;