# lib/api-zod/src/generated/types/orderStatusUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface OrderStatusUpdate {
  status: OrderStatus;
}