# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from './productCategory';
import type { ProductVariant } from './productVariant';
import type { StockStatus } from './stockStatus';

export interface Product {
  id: number;
  name: string;
  category: ProductCategory;
  price: number;
  description: string;
  photos: string[];
  variants: ProductVariant[];
  stockStatus: StockStatus;
  active: boolean;
  createdAt: Date;
}