# lib/api-zod/src/generated/types/productCategory.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ProductCategory = typeof ProductCategory[keyof typeof ProductCategory];


export const ProductCategory = {
  saree: 'saree',
  kurti: 'kurti',
  puja: 'puja',
  decor: 'decor',
} as const;