# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductCategory } from './productCategory';
import type { ProductVariant } from './productVariant';
import type { StockStatus } from './stockStatus';

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  category?: ProductCategory;
  /** @minimum 0 */
  price?: number;
  description?: string;
  photos?: string[];
  variants?: ProductVariant[];
  stockStatus?: StockStatus;
  active?: boolean;
}