# lib/api-zod/src/generated/types/stockStatus.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type StockStatus = typeof StockStatus[keyof typeof StockStatus];


export const StockStatus = {
  in_stock: 'in_stock',
  made_to_order: 'made_to_order',
} as const;